Structure, pattern table and tools.
Updates to remove the return so that patterns will queue - don't want to
restart the show everytime someone trips a sensor.
Added dybir - a new hypercolor that just picks a random color every
write.  Only supported by a couple of my patterns so far.
Added a whole lot of patterns to the main pattern table.